python_app/requirements.txt
Flask==2.2.5
google-cloud-firestore
google-cloud-tasks
google-cloud-secret-manager
google-auth>=2.35.0
gunicorn==21.2.0
requests==2.31.0
playwright==1.40.0
